requirements.txt
Django==2.0.3
django-filter==1.1.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
Markdown==2.6.11
PyJWT==1.6.1
pytz==2018.3
drf-yasg
psycopg2
django-fixture-magic